trogdor: Enable the host to self-report its sleep state
Define the CONFIG_POWER_TRACK_HOST_SLEEP_STATE, which enable the host to
self-report its sleep state. The handling function is not defined, so it
just saves the state, no further action.

BRANCH=None
BUG=b:148149387
TEST=Without this CL, will show HC 0xa9 error on suspend/resume:
[35.916083 HC 0xa9 err 1]
as the host sleep event (0xa9) is not handled.
With this CL, the HC 0xa9 error doesn't happen.

Change-Id: I68036d9ca611104eaf651b4d4cae15571713ac90
Signed-off-by: Wai-Hong Tam <waihong@google.com>