Make wait_for_ready available for all
wait_for_ready is a generic function which loops until bits in a
register are set. This patch move it to util.c to make it available
for all.

There are more places where the function is applicable but this CL
keeps the scope under chip/clock-stm32. There is no functionality
change.

BUG=none
BRANCH=none
TEST=buildall

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I796599344c1d86ab7144d1d6b434ec54cf1cc55d