TCPMv2: Don't swap away from Vconn source
This is a decision to put in the EC default policy long-term, but for
now just reject Vconn swaps when the DUT is sourcing in order to always
be able to communicate with cables.

Also correct tc_is_vconn_src() checks to reflect new behavior since the
function no longer returns less than 0.

BRANCH=None
BUG=b:148528713
TEST=on volteer with Startech dock, DUT does not swap Vconn away

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ia94255e1e5ac2476039b62b1f06a4295c309b1ed