power/mt8183: reset EC if EC has jumped and AP reboots
Resets EC if EC has jumped before and we want to boot AP.

In the normal flow, EC should stay at RW and reset itself
and AP should request EC for bootblock when start booting.

On some testing scenario, we might want to boot AP after RO->RW->RO
transition, and in such case, AP seems not in the correct off state
when S0->S5, so it won't send bootblock request to EC when we
try to turn on AP's rail and fails to boot.
To workaround this, we force EC reset if there's sysjump event happened.

BUG=b:161584167
TEST=boot AP; sysjump RO; apshutdown; powerbtn 1200; ensure AP boots
BRANCH=kukui

Change-Id: Ia4066f1764b83acf6835f344393dcad8c125d0e3
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>