flash: allow programming larger EC images
For some ECs, the EC image size is limited by the amount of code RAM
instead of 1/2 the total flash size. In this instance, there is unused
flash that can be used for single use data objects.

To support linking data objects into the unused flash area, increase the
region size that can be programmed for RW images.

Analysis of chips that are impacted by this change:
  NPCX             EC image size     New region size    Unused size
    NPCX5M5G:      96 KiB            128 KiB            32 KiB
    NPCX5M6G:      224 KiB           256 KiB            32 KiB
    NPCX7M6F:      192 KiB           256 KiB            64 KiB
    NPCX7M6FB:     192 KiB           256 KiB            64 KiB
    NPCX7M6FC:     192 KiB           256 KiB            64 KiB
    NPCX7M6G:      192 KiB           256 KiB            64 KiB
    NPCX7M7WC:     252 KiB           256 KiB            4 KiB
    NPCX7M7WB:     256 KiB           512 KiB            256 KiB
  MCHP
    mec17xx_2E00   188 KiB           256 KiB            68 KiB
    mec1701h       188 KiB           256 KiB            68 KiB
  MEC1322
    strago         104 KiB           128 KiB            24 KiB

Boards using other chips verified that CONFIG_RW_SIZE is the same as
CONFIG_EC_WRITABLE_STORAGE_SIZE.

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I84b9dc84568273e1ab1473e301d27ffd2b07ba7f