sc7180: Temporary hack to use the host sleep event to transit to S3
This is a temporary hack to use the host sleep event to transit to S3
instead of the AP_SUSPEND signal. The real AP_SUSPEND signal is ignored.
This hack is useful for people using the ToT kernel in which the
AP_SUSPEND doesn't work.

Must keep SPI interface active in S3, such that it can receive the
host sleep event. So enable the CONFIG_CHIPSET_RESUME_INIT_HOOK.
And notify RESUME_INIT hook on S5->S3 transition and SUSPEND_COMPLETE
hook on S3->S5 transition. This change may affect the power number a
bit, as the active SPI bus consumes more power.

This hack should be removed when AP_SUSPEND works on the ToT kernel.

BRANCH=None
BUG=b:148149387
TEST=In userspace, typing "powerd_dbug_suspend", EC transits to S3;
press a key to wake up, EC transits to S0.

Change-Id: I4a07ba4f2e824be207fb495f11b92f922af70c87
Signed-off-by: Wai-Hong Tam <waihong@google.com>