tcpmv2: Ensure that soft reset is sent within 15ms
Per TD.PD.SRC3.E24, the Soft Reset in response to an
extra Accept message should be with 15ms. We were sitting
at 18ms. When we request a reset, we should immediately process
the new state. (This does really make we want to try the change
to kick the state machine any time we call set_state to process the
next state transition quickly).

BRANCH=none
BUG=b:161836207
TEST=GRL's TD.PD.SRC3.E24 test now passes.

Signed-off-by: Jett Rink <jettrink@chromium.org>
Change-Id: I8cef1f2ed2a63a5262ee2d6205b87849dbff4f27