common: implement body_detection state machine
Implement body_detection state machine. Detect whether the user is
close to the devices or not.

Use the variance of acceleration from accelerometer to compute the
motion confidence. If motion confidence is high enough, set the motion
state to ON_BODY. If low enough, set it to OFF_BODY.

BRANCH=None
BUG=b:123434029
TEST=make buildall, flash the ec, watch the ec console log

Change-Id: If886b7f9fc505ef4db4d7abb08194b8f866121bb
Signed-off-by: Ching-Kang Yen <chingkang@chromium.org>