nct38xx: hard disable sinking when cc is set to open/open
zork's TCPC does not want to go into CC OPEN/OPEN when
it is powered solely by the power that it will be
cutting off. This should have caused the system to
brown out and reboot but the TCPC keeps enough of the
power connection to not allow a brown out but locks
up the rest of the system from continuing. By disabling
the SNKEN bit in CONTROL_OUT register, this condition
goes away and we will brown out on batteryless cold
reboot as expected.

BUG=b:162016100
BRANCH=none
TEST=ectool reboot_ec cold (without a battery attached)

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: Idb9597c978a9ba18f79ff158db6463af156707b5