Volteer: Add support for RT1715
Add support for replacing the TUSB422 with the RT1715. Since they are
pin-to-pin compatible and currently only used with a rework, make the
decision of which to use at runtime based on which I2C addresses are
available. The logic is to check if the TUSB422 is present; if not, use
the RT1715.

BUG=b:162617664
TEST=make buildall
TEST=check both TUSB422 functionality and RT1715 functionality
BRANCH=none

Change-Id: I8306f086bf030ddd7238532b1f12aa259cb72422
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>