ocpc: Add Istep parameter
Charger ICs differ in their ability to regulate down to a certain mA,
therefore, the OCPC code should take this ability into account and not
count it as an error if it's within the bounds of what the charger IC
is capable of regulating it to.

BUG=b:147440290
BRANCH=None
TEST=Enable on waddledee, flash, verify that once target current is
within an error bar of what the charger IC is capable of regulating
to, no error is reported.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: If94a9173e27d35be81b86589c530766bfe6431a2