sm5803: Add support for OCPC
This commit updates the SM5803 driver to add the required
functionality for supporting OCPC.  As this part does not have the
ability to measure ISYS, one needs to make sure that flag is set in
the OCPC chg_flags bitfield.

Currently, sourcing out the mainboard port while charging through an
auxiliary charger is not working.

BUG=b:155224387
BRANCH=None
TEST=Enable on waddledee, flash, verify DUT can charge thru the sub
board port.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I7b60dec51e503c83a24799d523eb75408f91ce86