usbc: correctly handle Get_Source_Cap as a sink
USB Power Delivery Specification Revision 3.0, version 2.0 section 6.3.7
states that a dual-role port shall respond to Get_Source_Cap with its
source capabilities, but this was incorrectly handled by responding with
a request for source capabilities.

Per section 8.3.3.18.10, implement the PE_DR_SNK_Give_Source_Cap state to
handle this correctly.

To support the new test, some helper functions for the fake PE are added
and the test code's copy of the PE state enum is updated to be in sync
with the real one.

BUG=b:161400825,b:161331630
TEST=New host test for this state, and verified on Dalboz that requesting a
     PRS via the EC console (`pd 1 swap power`) now sends source
     capabilities when the partner requests them.
BRANCH=None

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I87c27d406e0a3f57cf2c25fa583bee51155b6b12