Puff: Temperature sensor changes
The latest Puff board variants remove temperature sensor 2
and change the location of temperature sensor 1 to the SoC core.

Since one sensor has been removed, this change is backwards compatible
with previous boards (with the only exception that the location of the
sensor has changed, so the description is different). For previous
boards, the second temp sensor will be ignored.

BUG=b:162909373
TEST=Confirm that puff and variants only monitor one temp sensor.
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I0b03cbb24460a6569d36fbc6a0dab1c3c58dffbd