reef_it8320: update chip variant to dx version
This is development board with EC it8320bx, we gave 3 to intel
and 7 to google. Most of the development board, we had reworked
EC to it8320dx that is mass production version and flash size is
enough to define CONFIG_SYSTEM_UNLOCKED, so I update chip variant
to dx version, but this may not suit for those bx board.

BUG=none
BRANCH=none
TEST=can boot on board reef_it8320

Signed-off-by: Ruibin Chang <ruibin.chang@ite.com.tw>
Change-Id: I1f36ee15ffc515243863f73afb30a958613985ba