SM5803: Ignore low thershold interrupts when not throttled
If the temperature of the charger chip is hanging out around the lower
threshold for TINT, we may receive a fair number of interrupts as the
temperature fluctuates.  We only need to act on this interrupt if we
were previously throttling.

BRANCH=None
BUG=b:161546467
TEST=on drawlat, charged from the MB and noted that while the
temperature was near the lower threshold no console prints came through

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I34d08372eb1830dc4ffc9ce18342c170a19ea567