cbi: add SSFC CBI field
SSFC field will be leveraged to record what second source is used in the
DUT by probing components in the factory or RMA.

Firmware code should refer to this field to judge what driver should be
configured for a specific component. For example, the board code can
arrange what sensor driver should be set into motion_sensors array if
there are multiple sources of base or lid sensor.

As the definition of FW_CONFIG, it descripbe which "features" the
firmware code should enable or disable. For example, whether lid / base
sensors should be enabled or not but not care about what second source
is in this DUT.

BRANCH=none
BUG=b:163285687
TEST=call cbi-util to create the cbi image with SSFC and show created
content.

Change-Id: Icb4aa00ae47ab025198e7fd5edd6aab96a4bf53e
Signed-off-by: Marco Chen <marcochen@chromium.org>