Servo_v4: Change SBU detection logic
Servo V4 doesn't terminate the SBU lines, so they could be floating if
the DUT doesn't pull the SBU lines low. This can cause issues
with detection, since servov4 looks for the SBU lines being lower than
700mV to determine USB-C orientation. By USB2, ServoV4 should apply 15k
pulldowns as a USB2 host.

Change the detection logic to not requre D- to be <700mV, instead
determine orientation based on which SBU line is higher, and determine
connection based on one of the 2 SBU lines being above 2.5V

BUG=b:161762948, b:150886157, b:151487379
TEST=make BOARD=servo_v4
TEST=check that CCD works on DUT without pulldown resistors on SBU
BRANCH=none

Change-Id: I2b70a33888fadbd06f8a7a8dd990064df4d7a14c
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>