ServoV4: Change CCD detection flow
Currently, servoV4 doesn't apply terminations to SBU when searching for
a DUT. This means SBU can float which can break the SBU voltage
detection. Technically, the DUT should only need to pull up D+ and can
float D- since it is a FS USB2 device when in CCD mode.

Change the detection to connect SBU to the USB2 hub. The USB2 hub
will apply 15k pulldowns to the USB lines, which will pull SBU down and
fix the detection in the case where the DUT floats SBU.

BUG=b:161762948, b:150886157, b:151487379
TEST=make BOARD=servo_v4
TEST=check that CCD works on DUT without pulldown resistors on SBU
BRANCH=none

Change-Id: I9b6f620617fb0e270e35804704111afd3c606a3b
Signed-off-by: Eric Herrmann <eherrmann@chromium.org>