usbc: add GEN_N_S macro for NOT_SUPPORTED states
To enabled use of the IS_ENABLE macro we need to define some of the usb
state machine states as extern variables that will never be linked. This
allows the compiler to know about a state but will still fail the
linking phase if the config option is not defined properly.

This is just adding a convenience macro for this paradigm, not creating
the paradigm itself.

BRANCH=none
BUG=none
TEST=builds

Signed-off-by: Jett Rink <jettrink@chromium.org>
Change-Id: Ia19b5babd43d536f043745314481475cf25f6e01