sysrq: Ignore H1_EC_RECOVERY_BTN_ODL
On Puff, there are two signals from the recovery button. One is
from the button and the other is from H1.

The signal from H1 is masked by the power button and it looks
asserted only after the power button is released. This makes the
EC start the warm reset timer.

If the user doesn't release the recovery button (as instructed)
to enter the recovery mode for four seconds, the AP resets and
entry to the recovery mode fails.

This patch makes the EC ignore H1's signal as sysrq signal.

BUG=b:163443447
BRANCH=Puff
TEST=Recovery mode can be entered by keeping recovery button
pressed on Duffy.
TEST=Recovery button triggers sysrq.
TEST=Holding recovery button for 4 seconds triggers AP reset.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>
Change-Id: I8631d5dc0da8c95eed1d740b0683aa5de7848335