test: remove no_empty_state tests
When we use IS_ENABLED style code we may have empty state machine states
that are unused. The linker ensures that we never try to reference this
states. This does mean that extra states do cost ~20 bytes each. This
cost does seem worth it to keep the value of usb states stable.

BRANCH=none
BUG=none
TEST=buildall

Signed-off-by: Jett Rink <jettrink@chromium.org>
Change-Id: I44c1454bed91c9e28d89ebd6b75e9df684c41844