fingerprint: Stopgap to repair public builds
This repairs building bloonchipper and dartmonkey builds
when private lib is unavailable.

There are two issues this remedies:
1) Redefinition of sensor info macros, like FP_SENSOR_RES_X
2) Undefined reference to fp_sensor_maintenance when building
   bloonchipper/dartmonkey non-private image.

All of these issues stem from the way we handle code
when HAVE_FP_PRIVATE_DRIVER is set or faked by a test.
This is a fundamental issues that needs some thought and refactoring.

BRANCH=none
BUG=b:164174822,b:163109916
TEST=# With private
     time make buildall -j
TEST=# Without private
     time make buildall -j
TEST=# With private libs
     # Apply https://crrev.com/c/2357870
     ./util/compare_build.sh -b all

Signed-off-by: Craig Hesling <hesling@chromium.org>
Change-Id: If71e3376fb7a904b77f2fc9fc56c8e0daf54db3b