TCPMv2: Correct battery index detection
The battery index is located in the first byte after the extended
header, which is not passed up to the PE layer.  Additionally, the PRL
will start copying the message from byte 0 so correct the extended
message population.  Add condition to fill in a valid response if the
battery is not currently present, and correct the battery status
response initialization.

BRANCH=None
BUG=b:161837550
TEST=with GRL-C2, invalid battery index tests pass

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I65ebe56f197b9822a42463f8d9dfb42b34c1ffe1