Ezkinil: Workaround for reading unexpected value from isl9241 CONTROL1
Sometimes CONTROL1 was read 0xFF03 for unknown reason when the
state change from S0 to S3. If we do isl9241_discharge_on_ac
directly, we will write wrong value to enable OTG function (
bit11=1) that would make the system not charging.

Therefore, we need to check if the value read from CONTROL1 is
strange and modify it to default value.

BUG=b:163076059
BRANCH=none
TEST=none

Signed-off-by: Sue Chen <sue.chen@quanta.corp-partner.google.com>
Change-Id: Iec56dab838b0aa362c543cce74d64615faf40bfc