TCPMv2: SRC->SNK PR Swap enable AutoDischarge after Vbus
A debounce was added to allow for the TCPC to get a valid
Vbus but AutoDischargeDisconnect was enabled as soon as
we set for the debounce. Since Vbus had a good chance of
not being valid, we would get a FAULT from the TCPC
indicating an Auto Discharge Disconnect failure.

BUG=b:165797061
BRANCH=none
TEST=DUT-DUT PR_Swap

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I897d682e405d7f2fcda4b4a36a3c1b333e8f7913