tpcmv2: put USB mux into safe mode during suspend
Put the USB mux into safe mode during suspend operations. In case the
host wants to update firmware on any USB muxes, this ensures the mux is
in a safe state and also exits the mux from low power mode if it was
enabled.

BUG=b:162528867
BRANCH=none
TEST=make buildall
TEST=On Volteer, run "ectool pdcontrol suspend 1" and confirm that the
Burnside Bridge mux/retimer is brought out of reset.

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I99ee5a53ce7ffe46639d1fea9db23ccfd6e6e306