TCPMv2: re-enable GetSnkCap call for all
The problem was a discarded SendSrcCap message that
happened around the same time as the GetSnkCap.
GetSnkCap timedout and did what it was supposed to
do but the SendSrcCap locked up waiting for the
message to transfer before it started the timer
to allow it to time out.  Never would happen.

BUG=b:165822172
BRANCH=none
TEST=DUT-DUT PR_Swap

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I8854ac745d4489125c84b7e617d5fa1a94e70e51