SM5803: Add discharge on AC interface
In order to discharge on AC, the SM driver must disable sinking on its
chip.  This change also adds a loop to the common charger driver to set
discharge on AC for all chargers, since it's unlikely a caller of this
API would only this feature enabled for a single port.

BRANCH=None
BUG=b:164256610
TEST=on drawlat, run "ectool chargecontrol discharge" and "ectool
chargecontrol normal" with charger in C0 and observe discharging
followed by charging in "battery" output.  Repeat with C1

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ib23a93259a630016cb0f1384a6fc2b19b2baafb1