hack: temporarily work around TX_COMPLETE failure
After we removed the retry logic in the PRL layer, we uncovered an issue
that was being masked. On Delbin, the state machine is not receiving
that the PD messages (namely Request) is being sent successfully. We get
neither success or failure within 100ms. We still need to figure out and
fix this the root cause.

However there is an upcoming build for Delbin that we will have power on
issues if we don't add some kind of power work around. This patch will
not allow the USB PD stack on Delbin to work properly in all scenarios;
this patch only aims to ensure that we can get power to the device to
test other non-PD functionality.

This CL should be reverted within a two week time frame.

BRANCH=none
BUG=b:164154200
TEST=Delbin can power on

Signed-off-by: Jett Rink <jettrink@chromium.org>
Change-Id: If6dce35dfd78ee3a70e6216a7b6bf62d3ded5646