meep: enable 2nd TCPC source PS8755 for port 1
The original TCPC in the port 1 is PS8751 and this CL adds another
option of PS8755.

These two are all supported by ps8xxx.c and the board function
board_get_ps8xxx_product_id is implemented to judge the source by
checking the SSFC bits field in the CBI.

BUG=b:159082424, b:163922535
BRANCH=octopus
TEST=verify DUT with PS8755 or PS8751 in the sub-board can work
correctly in sink and source roles.

Signed-off-by: Marco Chen <marcochen@chromium.org>
Change-Id: I0cba58eb7b22c95aac1344f1b3a68ce5dac43ab0