octopus: create the probing of ps8xxx for SSFC of CBI
The CL hooks a probing into the init function of ps8xxx.c in order to
query which chip is in this DUT by asking the product_id from the chip.
Then the product_id is imported into helper function of SSFC for caching
the result.

These codes are gaurded by new CONFIG_FACTORY_SSFC_PROBE introduced in
this CL therefore the released firmware will be not with these probing
codes; partners need to enable this config for the local FW build
leveraged in the factory flow instead. And factory flow needs to update
formal firmware back to the DUT before finalization.

BRANCH=octopus
BUG=b:163922535
TEST=verify the EC build with CONFIG_FACTORY_SSFC_PROBE can detect
PS8751 or PS8755 in a DUT then update SSFC value accordingly.

Signed-off-by: Marco Chen <marcochen@chromium.org>
Change-Id: I8b4bb5faea0853ee40fe7ed87bc4786fd91642ed