TCPM: Correct VDO structures
When defined bit fields for a unit8_t cross byte boundaries, this can
cause unexpected packing.  In this case, all of the VDO unions were
actually of size 8, rather than size 4.  This change alters the fields
to be defined in a uint32_t, and adds build asserts to ensure the
unions remain the correct size.

BRANCH=None
BUG=None
TEST=make -j buildall

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ia3d585820dcb3e825c02f8732791b53fbdff96af