sm5803: Configure BFET alerts
The SM5803 has the capability to monitor to the power burnt across the
BFET.  There are two thresholds that can be configured: an early
threshold, and a fatal threshold in which the charger automatically
disables the BFET.

This commit configures the BFET alerts to fire at 1.5W for the early
threshold and 6.5W for the fatal threshold.

The full scale is 7.5W, the lsb is 7.5W/256 ~= 29.2mW.

BUG=b:159376384
BRANCH=None
TEST=Build and flash waddledee, charge from sub board, verify that no
alerts are seen.
TEST=Configure the thresholds significantly lower, charge from sub
board, verify that the alerts do fire and logging information is
printed on the EC console.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I17b3876bc5ed4b41d2378600a8b8bf639f9757ce