sc7180: Notify RESUME hooks in S3S0 if boot from off
After enabling the suspend hang detection, the RESUME hooks are deferred
to S0. When the kernel sets the host sleep event to flag the resume
completion, EC notifies the RESUME hooks. However, when it boots from an
off state, G3 or S5, the kernel won't set this host sleep event. Should
explicitly notify the RESUME hooks in the S3S0 state.

The change also renames the flag boot_from_g3 to boot_from_off and
carries it forward to S3S0; and renames the flag shutdown_from_s0 to
shutdown_from_on and carries it forward to S3S5.

BRANCH=None
BUG=b:148149387, b:167155164
TEST=Verified the RESUME hooks are called when boot from S5.

Change-Id: I48ee09ad66e53363e7a20d9602b37571177ac300
Signed-off-by: Wai-Hong Tam <waihong@google.com>