util: modify getversion to use proper timestamps.
The getversion.sh utility even when compiling the version string based
on the state of several git trees always uses the ec tree for
timestamps, be it the latest modified file if the tree is 'dirty' or
the last commit time if the tree is clean.

It should be using the latest time from all of the trees included in
the build.

BUG=none
TEST=verified operation for Cr50 with both main and secondary trees
     clean and dirty

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I72dc1d49ec997c789697b15f7d79fa9f4a8f8adc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2393101
Reviewed-by: Craig Hesling <hesling@chromium.org>
(cherry picked from commit 5ab4bd06261abf1204638c8ef877a9adb041d6e8)