TCPMv2: Restore the previous contract for non-EFS2
The previous CL:2271002 removes the BBRAM code and doesn't restore
the previous PD contract. Instead, it follows the PD recovery path.
But this only works for EFS2.

If EFS2 is not enabled, when sysjump to RW, that the PD partner
already has a PD contract, DUT won't receive any SourceCap message
in the PE_SNK_Wait_for_Capabilities state. It will trigger a hard
reset after timeout. This hard reset will cause the board brownout
and then a reboot loop.

This CL adds back the original code of restoring the previous
contract and PD soft reset logic. But only restore the contract
if it was a sink; otherwise, it follows the normal recovery path.
It sends PD soft reset on startup for the sysjump case (and the
EC warm reset case).

BRANCH=None
BUG=b:166307654,b:168134171
TEST=Tested on Pompom, without battery attached. EC warm reset and
sysjump to RW won't cause the board power-lost. Tested a 20V PD
charger, 5V non-PD charger, and through a Type-C hub with monitor
attached. Checked the data role is restored.

Change-Id: Icf8681bcc5c1e22cb378acea82142413c98230c9
Signed-off-by: Wai-Hong Tam <waihong@google.com>