ocpc: Only print CFET message if battery present
The OCPC algorithm will check to make sure that the battery CFET is
enabled before adjusting VSYS.  If it's not, it prints out an error
log message.  However, if the battery isn't present, the CFET can't be
enabled therefore this would lead to spammy output.

This commit just changes the log print to be emitted when the battery
is present AND the CFET is disabled.

BUG=none
BRANCH=none
TEST=`make -j buildall`

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ib0dcacd7736ca5f73c114c7cfa4446378086c58e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2393221
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Auto-Submit: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Diana Z <dzigterman@chromium.org>