util/flash_ec: do not pass 'unknown' or 'Uninitalized' S/N as args
Passing 'unknown' S/N as flashrom args might happen if we try to
`make BOARD=xyz flash_ec` a symlinked target 'xyz' (assume it points
'abc', and the servod is invoked by --board=abc). This is due to
the board name information in servod is coming the servod's argument,
however util/flash_ec would pass the $BOARD information to servod and
cause a wrong result while requesting the servo's serial number (by
which board connects to it).

This CL fix it by not passing the serial number for flashrom, and
let flashrom decides which servo to use.

BUG=b:168074868
TEST=sudo servod -b asurada;
     make BOARD=asurada flash_ec;
     make BOARD=hayato flash_ec
BRANCH=none

Change-Id: I6e190c483dd0c596a9ba294939bc5f85d363c016
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>