tcpmv2: Enable AutoDischarge on SRC->SNK PrSwap only if Vbus
AutoDischargeDisconnect should not be re-enabled during SRC->SNK
PR_Swap if Vbus is not present. This would happen only as an
error condition with the connected partner becoming unattached
during the swap.

Also, when we are SNK, we should disable/re-enable
AutoDischargeDisconnect when we hard reset because we are
shutting off sinking and Vbus will drop to Safe0V

BUG=b:168128121
BRANCH=none
TEST=PR_Swaps with bug mentioned dock

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I891857967559ecea0a7990626ef8fc3c6835797a