TCPMv2: Set the retimer to enabled alt mode on reboot
On chipset transition, the retimer is set to disconnect state. Since, the
port partner (DFP) doesn't send enter mode request again, this CL sets the
mux into previously entered alternate mode (if any) when the port
re-discovers the alternate modes

It also ensures that the discover queries are responded only in
S0 as the retimer turns on by main AP rail.

BUG=b:157163664
BRANCH=None
TEST=With port as UFP, able to enumerate Thunderbolt alt mode on reboot

Change-Id: I4cf7ef2112aacdb1a3f5a61ee2975a3c18ee5b71
Signed-off-by: Ayushee <ayushee.shah@intel.com>